+++ src/index.ts
@@ -21,6 +21,7 @@
 export * from "./types.js";
 export * from "./errors.js";
 export * from "./amount.js";
+export * from "./account.js";
 
 export * from "./discovery.js";
 export * from "./wallet.js";